feat(predicate): adds middleware to validate user permission to get predicate data
