+++ packages/api/src/middlewares/index.ts
@@ -1,3 +1,3 @@
 export * from './auth';
 export * from './handleErrors';
-export * from './permissions/index';
+export * from './permissions';
