+++ packages/api/src/modules/apiToken/routes.ts
@@ -1,5 +1,8 @@
 import { Router } from 'express';
-import { authMiddleware, predicatePermissionMiddleware } from '@src/middlewares';
+import {
+  authMiddleware,
+  predicateWorkspacePermissionMiddleware,
+} from '@src/middlewares';
 import { PermissionRoles } from '@src/models';
 import { APITokenController } from '@modules/apiToken/controller';
 import { handleResponse } from '@src/utils';
@@ -13,7 +16,7 @@ import {
 } from '@modules/apiToken/validations';
 
 const router = Router();
-const permissionMiddleware = predicatePermissionMiddleware({
+const permissionMiddleware = predicateWorkspacePermissionMiddleware({
   permissions: [
     PermissionRoles.OWNER,
     PermissionRoles.ADMIN,
